# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api Documentation jardin de cocagne",
        "version": "0.1"
    },
    "paths": {
...
        "/api/abonnement": {
            "get": {
                "tags": [
                    "Abonnements"
                ],
                "summary": "Get all subscriptions",
                "description": "Get a list of all subscriptions",
                "operationId": "getAllSubscriptions",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "Abonnements"
                ],
                "summary": "Add a new subscription",
                "description": "Add a new subscription for an adherent to a structure",
                "operationId": "addNewSubscription",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_structure",
                                    "id_adherent",
                                    "duree",
                                    "periodicite",
                                    "type_abonnement",
                                    "moyen_paiement"
                                ],
                                "properties": {
                                    "id_structure": {
                                        "type": "integer"
                                    },
                                    "id_adherent": {
                                        "type": "integer"
                                    },
                                    "duree": {
                                        "type": "integer"
                                    },
                                    "periodicite": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "type_abonnement": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "moyen_paiement": {
                                        "type": "string",
                                        "maxLength": 255
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Operation failed"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/abonnement/{id}": {
            "get": {
                "tags": [
                    "Abonnements"
                ],
                "summary": "Get a subscription by ID",
                "description": "Get details of a subscription by ID",
                "operationId": "getSubscriptionById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the subscription to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Subscription not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Abonnements"
                ],
                "summary": "Update a subscription",
                "description": "Update details of a subscription by ID",
                "operationId": "updateSubscription",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the subscription to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_structure",
                                    "id_adherent",
                                    "duree",
                                    "periodicite",
                                    "type_abonnement",
                                    "moyen_paiement"
                                ],
                                "properties": {
                                    "id_structure": {
                                        "type": "integer"
                                    },
                                    "id_adherent": {
                                        "type": "integer"
                                    },
                                    "duree": {
                                        "type": "integer"
                                    },
                                    "periodicite": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "type_abonnement": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "moyen_paiement": {
                                        "type": "string",
                                        "maxLength": 255
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Operation failed"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Abonnements"
                ],
                "summary": "Delete a subscription",
                "description": "Delete a subscription by ID",
                "operationId": "deleteSubscription",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the subscription to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Operation failed"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
...
}